+++ spec/models/work_package/work_package_watcher_spec.rb
@@ -40,6 +40,10 @@ describe WorkPackage do
                                               roles: [role]) }
     let(:work_package) { FactoryGirl.create(:work_package,
                                             project: project) }
+    let!(:user_allowed_to_view_work_packages) do
+      FactoryGirl.create(:user).tap { |user| project.add_member!(user, role) }
+    end
+
 
     context :recipients do
       let(:watcher) { Watcher.new(watchable: work_package,
@@ -66,6 +70,26 @@ describe WorkPackage do
         subject { work_package.watcher_recipients }
 
         it { should_not include(user.mail) }
+      end
+    end
+
+    context '#possible_watcher_users' do
+      it 'contains exactly those users who are allowed to view work packages' do
+        work_package.possible_watcher_users.should == [user_allowed_to_view_work_packages]
+      end
+    end
+
+    context 'notifications' do
+      let(:number_of_recipients) { (work_package.recipients | work_package.watcher_recipients).length }
+
+      before :each do
+        Delayed::Worker.delay_jobs = false
+      end
+
+      it 'sends one delayed mail notification for each possible watcher' do
+        UserMailer.stub_chain :issue_updated, :deliver
+        UserMailer.should_receive(:issue_updated).exactly(number_of_recipients).times
+        work_package.update_attributes :description => 'Any new description'
       end
     end
   end